requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
openai>=1.30.0
httpx>=0.27.0
pydantic>=2.7.0
loguru>=0.7.2
jsonschema>=4.22.0
groq==0.11.0